backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
sqlalchemy[asyncio]==2.0.41
asyncpg==0.31.0
alembic==1.15.2
pydantic==2.11.3
pydantic-settings==2.9.1
httpx==0.28.1
google-genai==1.14.0
groq==0.25.0
python-dotenv==1.1.0
